[22/09/2023] Separated hooks for search
